# pkg/web/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for querying information about Assertoor tests",
        "title": "Assertoor API Documentation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/test/{testId}": {
            "get": {
                "description": "Returns the test definition with given ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Get test definition by test ID",
                "operationId": "getTest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the test definition to get details for",
                        "name": "testId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web_api.GetTestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/test_run/{runId}": {
            "get": {
                "description": "Returns the run details with given ID. Includes a summary and a list of task with limited details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TestRun"
                ],
                "summary": "Get test run by run ID",
                "operationId": "getTestRun",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the test run to get details for",
                        "name": "runId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web_api.GetTestRunResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/test_runs": {
            "get": {
                "description": "Returns a list of all test runs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TestRun"
                ],
                "summary": "Get list of test runs",
                "operationId": "getTestRuns",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Return test runs for this test ID only",
                        "name": "test_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web_api.GetTestRunsResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/tests": {
            "get": {
                "description": "Returns the list of test definitions. These test definitions can be used to create new test runs and are supplied via the assertoor configuration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Get list of test definitions",
                "operationId": "getTests",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web_api.GetTestsResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/web_api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.TestSchedule": {
            "type": "object",
            "properties": {
                "cron": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "skipQueue": {
                    "type": "boolean"
                },
                "startup": {
                    "type": "boolean"
                }
            }
        },
        "types.TestStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "success",
                "failure",
                "skipped",
                "aborted"
            ],
            "x-enum-varnames": [
                "TestStatusPending",
                "TestStatusRunning",
                "TestStatusSuccess",
                "TestStatusFailure",
                "TestStatusSkipped",
                "TestStatusAborted"
            ]
        },
        "web_api.GetTestResponse": {
            "type": "object",
            "properties": {
                "basePath": {
                    "type": "string"
                },
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "configVars": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/types.TestSchedule"
                },
                "source": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
...
        "web_api.GetTestRunResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "run_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.TestStatus"
                },
                "stop_time": {
                    "type": "integer"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web_api.GetTestRunTask"
                    }
                },
                "test_id": {
                    "type": "string"
                }
            }
        },
        "web_api.GetTestRunTask": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "index": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_index": {
                    "type": "integer"
                },
                "result": {
                    "type": "string"
                },
                "result_error": {
                    "type": "string"
                },
                "result_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web_api.GetTestRunTaskResult"
                    }
                },
                "runtime": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "integer"
                },
                "started": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "stop_time": {
                    "type": "integer"
                },
                "timeout": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "web_api.GetTestRunTaskResult": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "web_api.GetTestRunsResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "run_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.TestStatus"
                },
                "stop_time": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "string"
                }
            }
        },
        "web_api.GetTestsResponse": {
            "type": "object",
            "properties": {
                "basePath": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "web_api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    },
...
}